Remove Pi-hole from the iot doc, as they state:
"During the development of v6 – we took a decision to remove the temperature from the web interface. It was never consistently displayed on every user’s interface – sometimes we couldn’t get the correct sensor reading, and sometimes we couldn’t get a sensor reading at all. We would find ourselves spending a lot of time debugging issues with sensors for very little gain."